Mois 61 id 004 edit house information (#26)
* feat: profile page + change username

* feat: userStats on profilepage

* feat: display user booking stats profile

* feat: user profile image upload + user stats fix

* test: added test for user profile page

* fix: remove settings from navbar

* Update app/components/userStats.tsx


* fix: profile reactive+minor fixes

* feat: add owner field

* feat: alert context for leaving and deleting houses

* feat: delegate, leave or delete by role

* fix: remove cypress commands

* feat: edit house information

* feat: edit house for owner role+ house image upload

* fix: added firebasestorage domain for images

* fix: fix image alt

+++ app/pages/household/[houseId]/index.tsx
@@ -1,7 +1,10 @@
-import React, { useContext, useRef } from "react";
+import React, { useContext, useState, useEffect, useRef } from "react";
 
 import { AuthContext } from "../../../context/AuthContext";
+import AlertContext from "../../../context/alertProvider";
+
 import { useRouter } from "next/router";
+import Image from "next/image";
 
 import {
   addDoc,
@@ -10,8 +13,7 @@ import {
   arrayRemove,
   doc,
   collection,
-  where,
-  query,
+  getDoc,
 } from "firebase/firestore";
 import { db } from "../../../firebase/clientApp";
 import { useCollection, useDocument } from "react-firebase-hooks/firestore";
@@ -21,25 +23,64 @@ import NewUtilityForm from "../../../components/newUtilityForm";
 import HouseholdCalendar from "../../../components/householdCalendar";
 import HouseholdMembers from "../../../components/household/householdMembers";
 import InviteCard from "../../../components/userInvites/inviteCard";
+import Alert from "../../../components/alert";
+
+import HouseholdEditModal, {
+  householdEditModalId,
+} from "../../../components/modals/householdEditModal";
 
 const Household = () => {
   const router = useRouter();
   const { houseId }: any = router.query;
   const authContext = useContext(AuthContext);
+  const alert = useContext(AlertContext);
   const currentUser = authContext?.userData;
+
   const householdRef = doc(db, "household", houseId);
   const amenityRef = collection(householdRef, "amenity");
-  //const userRef = doc(householdRef, "users");
-  const bookingsRef = collection(db, "booking");
 
-  const modalCheckboxRef = useRef<HTMLInputElement>(null);
+  const createUtilityModalRef = useRef<HTMLInputElement>(null);
+  0;
+  const deleteModalCheckboxRef = useRef<HTMLInputElement>(null);
 
-  const [value, loading, error] = useDocument(householdRef);
-  const [bookingValue, bookingLoading, bookingError] = useCollection(
-    query(bookingsRef, where("householdId", "==", houseId))
-  );
+  const [household, loading, error] = useDocument(householdRef);
   const [amenityValue, amenityLoading, amenityError] =
     useCollection(amenityRef);
+
+  const [householdMembers, setHouseholdMembers] = useState<any[]>();
+
+  useEffect(() => {
+    let active = true;
+    loadMembers();
+    return () => {
+      active = false;
+    };
+
+    async function loadMembers() {
+      if (household) {
+        const allMembers = await Promise.all(
+          household?.data()?.users.map(async (memberId: any) => {
+            const memberRef = doc(db, "users", memberId);
+            const memberName = await getDoc(memberRef).then((data) => {
+              return data.data()?.data.displayName;
+            });
+            return [memberId, memberName];
+          })
+        );
+        const withoutUser = allMembers.filter((e: any, i: any, a: any) => {
+          return e[0] !== currentUser?.userId;
+        });
+        if (!active) {
+          return;
+        }
+        setHouseholdMembers(withoutUser);
+      }
+    }
+  }, []);
+
+  const isOwner = (): boolean => {
+    return household?.data()?.owner === currentUser?.userId;
+  };
 
   const handleCreateAmenity = async (event: any) => {
     event.preventDefault();
@@ -50,28 +91,64 @@ const Household = () => {
       type: type.value,
       desc: desc.value,
     });
-    if (modalCheckboxRef.current !== null) {
-      modalCheckboxRef.current.checked = !modalCheckboxRef.current.checked;
+    if (createUtilityModalRef.current !== null) {
+      createUtilityModalRef.current.checked =
+        !createUtilityModalRef.current.checked;
     }
   };
 
+  // Delete a household if the user is the owner of the household
   const handleDeleteHouse = async (event: any) => {
+    if (!isOwner()) {
+      alert?.error("Need owner permission to delete household!", 5);
+      return;
+    }
     router.push("/");
     const docRef = await deleteDoc(doc(db, "household", householdRef.id));
     console.log("House deleted");
+
+    const message = `Successfully deleted ${household?.data()?.name}`;
+    alert?.success(message, 5);
   };
 
+  /** Leave a household.
+   * Cannot leave if user is the owner of Household (Need to delegate ownership).
+   * This automatically means that the last member (the owner) cannot leave the household.
+   * */
   const handleLeaveHouse = async (event: any) => {
-    router.push("/");
     const currentUserId = currentUser?.userId;
     const docRef = doc(db, "household", householdRef.id);
+    router.push("/");
     try {
       await updateDoc(docRef, { users: arrayRemove(currentUserId) });
     } catch (error) {
       console.log("error");
-      alert(error);
+      alert?.error(`An error has occured.`, 5);
     }
     console.log("House left");
+
+    const message = `Successfully left ${household?.data()?.name}`;
+    alert?.success(message, 5);
+  };
+
+  /**
+   * Handler that delegates the ownership of the household to another user.
+   */
+  const handleDelegateOwner = async (event: any) => {
+    event.preventDefault();
+    const newOwnerId = event.target.elements.newOwner.value.split(",")[0];
+    const newOwnerName = event.target.elements.newOwner.value.split(",")[1];
+
+    const updateOwner = await updateDoc(householdRef, {
+      owner: newOwnerId,
+    }).then(() => {
+      alert?.success(`Delegated ownership to ${newOwnerName}`, 5);
+    });
+
+    if (deleteModalCheckboxRef.current !== null) {
+      deleteModalCheckboxRef.current.checked =
+        !deleteModalCheckboxRef.current.checked;
+    }
   };
 
   return (
@@ -101,7 +178,7 @@ const Household = () => {
           <input
             type="checkbox"
             id="delete-house-modal"
-            ref={modalCheckboxRef}
+            ref={deleteModalCheckboxRef}
             className="modal-toggle"
           />
           <div className="modal">
@@ -112,23 +189,102 @@ const Household = () => {
               >
                 ✕
               </label>
-              <h3 className="text-lg font-bold">
-                Are you sure you want to delete the house?
-              </h3>
-              <p className="py-4">
-                This action will delete the house for all users.
-              </p>
-              <input
-                className="btn"
-                onClick={handleLeaveHouse}
-                value="leave house instead"
-              ></input>
+              <h3 className="text-lg font-bold">Household Settings</h3>
+              {!isOwner() && (
+                <div className="flex flex-col mt-5">
+                  <h3 className="text-xl font-bold">Leave House</h3>
+                  <p>
+                    You&apos;ll need a new invitation to join this house again!
+                  </p>
+                  <input
+                    className="btn btn-error rounded-lg place-self-center mt-5"
+                    onClick={handleLeaveHouse}
+                    value="leave house"
+                  ></input>
+                </div>
+              )}
               <br></br>
-              <input
-                className="btn"
-                onClick={handleDeleteHouse}
-                value="delete house"
-              ></input>
+              {isOwner() && (
+                <div className="flex flex-col">
+                  <form action="#" onSubmit={handleDelegateOwner}>
+                    <h3 className="text-xl font-bold">Delegate Ownership</h3>
+                    <p>
+                      Choose someone to handover the ownership of this UHouse!
+                    </p>
+                    <br />
+                    <div className="flex gap-x-5">
+                      {householdMembers && householdMembers.length > 0 ? (
+                        <>
+                          <select
+                            name="newOwner"
+                            id="newOwner"
+                            className="select select-primary w-full max-w-xs"
+                          >
+                            <option disabled selected>
+                              Select the new owner!
+                            </option>
+                            {householdMembers.map((member: any) => {
+                              return (
+                                <option key={member[0]} value={member}>
+                                  {member[1]}
+                                </option>
+                              );
+                            })}
+                          </select>
+                          <input
+                            type="submit"
+                            className="btn btn-warning"
+                            value="Delegate"
+                          ></input>
+                        </>
+                      ) : (
+                        <>
+                          <select
+                            name="newOwner"
+                            id="newOwner"
+                            className="select w-full max-w-xs"
+                            disabled
+                          >
+                            <option>You&apos;re the only one left!</option>
+                          </select>
+                          <input
+                            type="submit"
+                            className="btn btn-primary"
+                            value="Delegate"
+                            disabled
+                          ></input>
+                        </>
+                      )}
+                    </div>
+                  </form>
+                  <br />
+                  <h3 className="text-xl font-bold">Delete Household</h3>
+                  <div className="alert alert-warning shadow-lg">
+                    <div>
+                      <svg
+                        xmlns="http://www.w3.org/2000/svg"
+                        className="stroke-current flex-shrink-0 h-6 w-6"
+                        fill="none"
+                        viewBox="0 0 24 24"
+                      >
+                        <path
+                          strokeLinecap="round"
+                          strokeWidth="2"
+                          d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
+                        />
+                      </svg>
+                      <span>Warning: You cannot undo this action!</span>
+                    </div>
+                  </div>
+                  <button
+                    className="mt-5 btn btn-error rounded-lg place-self-center"
+                    onClick={handleDeleteHouse}
+                    id="deleteHouseholdButton"
+                  >
+                    delete house
+                  </button>
+                </div>
+              )}
             </div>
           </div>
           <label
@@ -141,7 +297,7 @@ const Household = () => {
           <input
             type="checkbox"
             id="new-utility-modal"
-            ref={modalCheckboxRef}
+            ref={createUtilityModalRef}
             className="modal-toggle"
           />
           <div className="modal">
@@ -161,11 +317,26 @@ const Household = () => {
           </div>
         </div>
         <div className="basis-1/2">
-          <h1 className="text-center font-black text-3xl mb-2">
-            {" "}
-            Household {value?.data()?.name}
-            <div className="dropdown dropdown-right">
-              <label tabIndex={0} className="btn btn-circle btn-ghost btn-xs">
+          <div className="gap-x-5 flex items-end justify-center">
+            <Image
+              className="rounded-xl avatar"
+              src={
+                household?.data()?.photoURL ||
+                "https://placeimg.com/80/80/house"
+              }
+              width={90}
+              height={90}
+              alt="household pic"
+            />
+            <h1 className="text-center font-black text-3xl mb-5">
+              Household {household?.data()?.name}
+            </h1>
+            <div className="dropdown dropdown-right mb-5">
+              <label
+                tabIndex={0}
+                id="householdMenu"
+                className="btn btn-circle btn-ghost btn-xs"
+              >
                 <svg
                   xmlns="http://www.w3.org/2000/svg"
                   fill="none"
@@ -185,21 +356,28 @@ const Household = () => {
                 tabIndex={0}
                 className="mt-3 p-2 shadow menu menu-compact dropdown-content bg-base-100 rounded-box w-52"
               >
+                {isOwner() && (
+                  <li>
+                    <label
+                      id="householdEditSettings"
+                      htmlFor={householdEditModalId}
+                    >
+                      Edit
+                    </label>
+                  </li>
+                )}
                 <li>
-                  <a className="justify-between">Edit</a>
-                </li>
-                <li>
-                  <label htmlFor="delete-house-modal" className="text-red-500">
-                    Delete
+                  <label
+                    id="householdDeleteSettings"
+                    htmlFor="delete-house-modal"
+                    className="text-red-500"
+                  >
+                    Delete/Delegate
                   </label>
-                  {/*
-                  <label htmlFor="leave-house-modal" className="text-red-500">
-                    Leave
-                    </label>*/}
                 </li>
               </ul>
             </div>
-          </h1>
+          </div>
           <HouseholdCalendar houseId={houseId}></HouseholdCalendar>
         </div>
         <div className="basis-1/4">
@@ -207,7 +385,9 @@ const Household = () => {
           <HouseholdMembers houseId={houseId} />
           <InviteCard houseId={houseId} />
         </div>
+        <Alert />
       </div>
+      {isOwner() && <HouseholdEditModal houseId={houseId} />}
     </>
   );
 };
